Move transaction type extraction to its own class
