+++ src/main/java/tools/sctrade/companion/domain/commodity/CommoditySubmissionFactory.java
@@ -1,6 +1,5 @@
 package tools.sctrade.companion.domain.commodity;
 
-import java.awt.Rectangle;
 import java.awt.image.BufferedImage;
 import java.time.Instant;
 import java.util.ArrayList;
@@ -8,16 +7,13 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.TreeMap;
-import java.util.concurrent.ConcurrentSkipListMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import tools.sctrade.companion.domain.LocationRepository;
 import tools.sctrade.companion.domain.image.ImageManipulation;
-import tools.sctrade.companion.domain.image.ImageType;
 import tools.sctrade.companion.domain.image.ImageWriter;
 import tools.sctrade.companion.domain.image.manipulations.CommodityKioskTextThreshold;
 import tools.sctrade.companion.domain.image.manipulations.ConvertToGreyscale;
@@ -28,23 +24,23 @@ import tools.sctrade.companion.domain.ocr.LocatedColumn;
 import tools.sctrade.companion.domain.ocr.LocatedFragment;
 import tools.sctrade.companion.domain.ocr.Ocr;
 import tools.sctrade.companion.domain.ocr.OcrResult;
+import tools.sctrade.companion.domain.ocr.OcrUtil;
 import tools.sctrade.companion.domain.user.UserService;
 import tools.sctrade.companion.exceptions.LocationNotFoundException;
 import tools.sctrade.companion.exceptions.NoCloseStringException;
 import tools.sctrade.companion.exceptions.NoListingsException;
 import tools.sctrade.companion.exceptions.NotEnoughColumnsException;
 import tools.sctrade.companion.utils.HashUtil;
-import tools.sctrade.companion.utils.ImageUtil;
 import tools.sctrade.companion.utils.LocalizationUtil;
 import tools.sctrade.companion.utils.StringUtil;
 import tools.sctrade.companion.utils.TimeUtil;
 
 public class CommoditySubmissionFactory {
-  private static final String SHOP_INVENTORY = "shop inventory";
   private static final String YOUR_INVENTORIES = "your inventories";
 
   private final Logger logger = LoggerFactory.getLogger(CommoditySubmissionFactory.class);
 
+  private TransactionTypeExtractor transactionTypeExtractor;
   private UserService userService;
   private NotificationService notificationService;
   private CommodityRepository commodityRepository;
@@ -56,6 +52,7 @@ public class CommoditySubmissionFactory {
   public CommoditySubmissionFactory(UserService userService,
       NotificationService notificationService, CommodityRepository commodityRepository,
       LocationRepository locationRepository, ImageWriter imageWriter) {
+    this.transactionTypeExtractor = new TransactionTypeExtractor(imageWriter);
     this.userService = userService;
     this.notificationService = notificationService;
     this.commodityRepository = commodityRepository;
@@ -71,7 +68,8 @@ public class CommoditySubmissionFactory {
       OcrResult listingsResult = listingsOcr.get().read(screenCapture);
       var rawListings = buildRawListings(listingsResult);
       logger.debug("Read {} listings", rawListings.size());
-      var transactionType = extractTransactionType(screenCapture, listingsResult);
+      TransactionType transactionType =
+          transactionTypeExtractor.extract(screenCapture, listingsResult);
 
       logger.debug("Reading location...");
       OcrResult locationResult = locationOcr.get().read(screenCapture);
@@ -156,53 +154,10 @@ public class CommoditySubmissionFactory {
     return rawListings;
   }
 
-  private TransactionType extractTransactionType(BufferedImage screenCapture, OcrResult result) {
-    Rectangle shopInv = getShopInventoryRectangle(result);
-    screenCapture = ImageUtil.crop(screenCapture, new Rectangle((screenCapture.getWidth() / 2), 0,
-        (screenCapture.getWidth() - (screenCapture.getWidth() / 2)), screenCapture.getHeight()));
-    screenCapture = ImageUtil.crop(screenCapture,
-        new Rectangle((int) shopInv.getMinX(), (int) (shopInv.getMinY() - shopInv.getHeight()),
-            (int) (shopInv.getWidth() * 3), (int) (shopInv.getHeight() * 4)));
-    screenCapture = ImageUtil.makeGreyscaleCopy(screenCapture);
-
-    var image = ImageUtil.makeCopy(screenCapture);
-    image = ImageUtil.applyOtsuBinarization(image);
-    imageWriter.write(image, ImageType.BUTTONS);
-    var boundingBoxes = ImageUtil.findBoundingBoxes(image);
-
-    boundingBoxes
-        .sort((n, m) -> Double.compare(m.getWidth() * m.getHeight(), n.getWidth() * n.getHeight()));
-    var boundingBoxesOrderedByX = boundingBoxes.subList(0, 2);
-    boundingBoxesOrderedByX.sort((n, m) -> Double.compare(n.getMinX(), m.getMinX()));
-    Rectangle buyRectangle = boundingBoxesOrderedByX.get(0);
-    Rectangle sellRectangle = boundingBoxesOrderedByX.get(1);
-
-    var buyImage = ImageUtil.crop(screenCapture, buyRectangle);
-    var buyRectangleColor = ImageUtil.calculateAverageColor(buyImage);
-    var buyRectangleLuminance = buyRectangleColor.getRed();
-    imageWriter.write(buyImage, ImageType.BUY_BUTTON);
-
-    var sellImage = ImageUtil.crop(screenCapture, sellRectangle);
-    var sellRectangleColor = ImageUtil.calculateAverageColor(sellImage);
-    var sellRectangleLuminance = sellRectangleColor.getRed();
-    imageWriter.write(sellImage, ImageType.SELL_BUTTON);
-
-    return (buyRectangleLuminance > sellRectangleLuminance) ? TransactionType.SELLS
-        : TransactionType.BUYS;
-  }
-
-  private Rectangle getShopInventoryRectangle(OcrResult result) {
-    var fragments = result.getColumns().parallelStream()
-        .flatMap(n -> n.getFragments().parallelStream()).toList();
-    LocatedFragment shopInventoryFragment = findFragmentClosestTo(fragments, SHOP_INVENTORY);
-
-    return shopInventoryFragment.getBoundingBox();
-  }
-
   private String extractLocation(OcrResult result) {
     List<LocatedFragment> fragments =
         result.getColumns().stream().flatMap(n -> n.getFragments().stream()).toList();
-    var yourInventoriesFragment = findFragmentClosestTo(fragments, YOUR_INVENTORIES);
+    var yourInventoriesFragment = OcrUtil.findFragmentClosestTo(fragments, YOUR_INVENTORIES);
 
     // Return the fragment that follows "your inventories"
     var it = fragments.iterator();
@@ -227,19 +182,6 @@ public class CommoditySubmissionFactory {
     }
 
     throw new LocationNotFoundException(fragments);
-  }
-
-  private LocatedFragment findFragmentClosestTo(Collection<LocatedFragment> fragments,
-      String string) {
-    Map<Integer, LocatedFragment> fragmentsByDistanceToTarget = new ConcurrentSkipListMap<>();
-    fragments.parallelStream().forEach(n -> fragmentsByDistanceToTarget
-        .put(StringUtil.calculateLevenshteinDistance(n.getText(), string), n));
-
-    if (fragmentsByDistanceToTarget.keySet().iterator().next() > (string.length() / 2)) {
-      throw new NoCloseStringException(string);
-    }
-
-    return fragmentsByDistanceToTarget.values().iterator().next();
   }
 
   private Iterator<LocatedColumn> getColumnIteratorOrderedByLineCount(List<LocatedColumn> columns) {
